refactor: movidos componentes do RealEstate para o módulo Organization
