+++ modules/Organization/Models/OrganizationAddress.php
@@ -1,0 +1,105 @@
+<?php
+
+/**
+ * @author      Luan Silva
+ * @copyright   2025 The Dev Kitchen (https://www.thedevkitchen.com.br)
+ * @license     https://www.thedevkitchen.com.br  Copyright
+ */
+
+declare(strict_types=1);
+
+namespace Modules\Organization\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\SoftDeletes;
+
+/**
+ * Modelo para endereços de organizações
+ */
+class OrganizationAddress extends Model
+{
+    use HasFactory, SoftDeletes;
+
+    /**
+     * A tabela associada ao modelo.
+     *
+     * @var string
+     */
+    protected $table = 'organization_addresses';
+
+    /**
+     * Os atributos que são atribuíveis em massa.
+     *
+     * @var array<string>
+     */
+    protected $fillable = [
+        'organization_id',
+        'organization_type',
+        'type',
+        'street',
+        'number',
+        'complement',
+        'neighborhood',
+        'city',
+        'state',
+        'zip_code',
+        'country',
+        'active',
+    ];
+
+    /**
+     * Os atributos que devem ser convertidos.
+     *
+     * @var array<string, string>
+     */
+    protected $casts = [
+        'active' => 'boolean',
+        'created_at' => 'datetime',
+        'updated_at' => 'datetime',
+    ];
+
+    /**
+     * Obtém a organização a que este endereço pertence
+     *
+     * @return BelongsTo
+     */
+    public function organization(): BelongsTo
+    {
+        return $this->belongsTo(Organization::class, 'organization_id');
+    }
+
+    /**
+     * Escopo para filtrar apenas endereços ativos
+     *
+     * @param \Illuminate\Database\Eloquent\Builder $query
+     * @return \Illuminate\Database\Eloquent\Builder
+     */
+    public function scopeActive($query)
+    {
+        return $query->where('active', true);
+    }
+
+    /**
+     * Escopo para filtrar apenas endereços de matriz
+     *
+     * @param \Illuminate\Database\Eloquent\Builder $query
+     * @return \Illuminate\Database\Eloquent\Builder
+     */
+    public function scopeHeadquarters($query)
+    {
+        return $query->where('type', 'headquarters');
+    }
+
+    /**
+     * Escopo para filtrar apenas endereços de filiais
+     *
+     * @param \Illuminate\Database\Eloquent\Builder $query
+     * @return \Illuminate\Database\Eloquent\Builder
+     */
+    public function scopeBranches($query)
+    {
+        return $query->where('type', 'branch');
+    }
+}
