--- modules/RealEstate/GraphQL/Mutations/DeleteRealEstate.php
@@ -1,50 +1,0 @@
-<?php
-
-/**
- * @author      Luan Silva
- * @copyright   2025 The Dev Kitchen (https://www.thedevkitchen.com.br)
- * @license     https://www.thedevkitchen.com.br  Copyright
- */
-
-declare(strict_types=1);
-
-namespace Modules\RealEstate\GraphQL\Mutations;
-
-use GraphQL\Type\Definition\ResolveInfo;
-use Modules\RealEstate\Models\RealEstate;
-use Modules\RealEstate\Services\RealEstateService;
-use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
-
-class DeleteRealEstate
-{
-    /**
-     * @var RealEstateService
-     */
-    private RealEstateService $_realEstateService;
-
-    /**
-     * Constructor.
-     *
-     * @param RealEstateService $realEstateService
-     */
-    public function __construct(RealEstateService $realEstateService)
-    {
-        $this->_realEstateService = $realEstateService;
-    }
-
-    /**
-     * Delete a real estate agency.
-     *
-     * @param mixed $rootValue
-     * @param array $args
-     * @param GraphQLContext $context
-     * @param ResolveInfo $resolveInfo
-     * @return RealEstate
-     */
-    public function __invoke($rootValue, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): RealEstate
-    {
-        // Extract ID and delegate to service layer for business logic
-        $id = (int) $args['id'];
-        return $this->_realEstateService->deleteRealEstate($id);
-    }
-}
