--- modules/RealEstate/GraphQL/Queries/RealEstates.php
@@ -1,81 +1,0 @@
-<?php
-
-/**
- * @author      Luan Silva
- * @copyright   2025 The Dev Kitchen (https://www.thedevkitchen.com.br)
- * @license     https://www.thedevkitchen.com.br  Copyright
- */
-
-declare(strict_types=1);
-
-namespace Modules\RealEstate\GraphQL\Queries;
-
-use GraphQL\Type\Definition\ResolveInfo;
-use Modules\RealEstate\Models\RealEstate;
-use Modules\RealEstate\Services\RealEstateService;
-use Modules\UserManagement\Database\Seeders\RolesSeeder;
-use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
-
-class RealEstates
-{
-    private RealEstateService $_realEstateService;
-
-    /**
-     * Constructor.
-     */
-    public function __construct(RealEstateService $realEstateService)
-    {
-        $this->_realEstateService = $realEstateService;
-    }
-
-    /**
-     * Return a paginated list of real estate agencies.
-     */
-    public function __invoke($rootValue, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): array
-    {
-        // Authorize access to real estate data
-        $user = $this->_realEstateService->authorizeRealEstateAccess();
-
-        $query = RealEstate::query();
-
-        // Apply multi-tenant filtering for non-admin users
-        if ($user->role && $user->role->name !== RolesSeeder::ROLE_SUPER_ADMIN) {
-            // Filter by tenant for regular users
-            if ($user->tenant_id) {
-                $query->where('tenant_id', $user->tenant_id);
-            }
-        }
-
-        $first = $args['first'] ?? 10;
-        $page = $args['page'] ?? 1;
-
-        // Apply ordering if specified
-        if (isset($args['orderBy']) && is_array($args['orderBy'])) {
-            foreach ($args['orderBy'] as $orderBy) {
-                $direction = $orderBy['order'] === 'DESC' ? 'desc' : 'asc';
-                $query->orderBy($orderBy['column'], $direction);
-            }
-        } else {
-            // Default ordering
-            $query->orderBy('name', 'asc');
-        }
-
-        // Get paginated results
-        $paginatedRealEstates = $query->paginate($first, ['*'], 'page', $page);
-
-        // Return data in the format expected by the GraphQL schema
-        return [
-            'data' => $paginatedRealEstates->items(),
-            'paginatorInfo' => [
-                'count' => $paginatedRealEstates->count(),
-                'currentPage' => $paginatedRealEstates->currentPage(),
-                'firstItem' => $paginatedRealEstates->firstItem(),
-                'hasMorePages' => $paginatedRealEstates->hasMorePages(),
-                'lastItem' => $paginatedRealEstates->lastItem(),
-                'lastPage' => $paginatedRealEstates->lastPage(),
-                'perPage' => $paginatedRealEstates->perPage(),
-                'total' => $paginatedRealEstates->total(),
-            ],
-        ];
-    }
-}
